+++ esibd/devices/omnicontrol/omnicontrol.py
@@ -1,11 +1,12 @@
 # pylint: disable=[missing-module-docstring]  # see class docstrings
+from enum import Enum
 from typing import cast
 
 import numpy as np
-import pfeiffer_vacuum_protocol as pvp
 import serial
+from PyQt6.QtWidgets import QPushButton
 
-from esibd.core import PARAMETERTYPE, PLUGINTYPE, PRINT, Channel, DeviceController, Parameter, parameterDict
+from esibd.core import PARAMETERTYPE, PLUGINTYPE, PRINT, Channel, DeviceController, Parameter, ToolButton, parameterDict
 from esibd.plugins import Device, Plugin
 
 
@@ -15,7 +16,12 @@ def providePlugins() -> 'list[type[Plugin]]':
 
 
 class OMNICONTROL(Device):
-    """Reads pressure values form an Pfeiffer Omnicontrol using the Pfeiffer Vacuum Protocol."""
+    """Reads pressure values and controls turbo pumps from Pfeiffer.
+
+    Can be used with and without Omnicontrol hardware.
+    Access to frequently uses pump parameters is provided by the controller and convenience methods of the channel.
+    Additional parameters can be added following the same pattern if needed.
+    """
 
     name = 'OMNICONTROL'
     version = '1.0'
@@ -24,15 +30,16 @@ class OMNICONTROL(Device):
     unit = 'mbar'
     iconFile = 'pfeiffer_omni.png'
     logY = True
-    channels: 'list[PressureChannel]'
+    channels: 'list[OmniChannel]'
+    controller: 'OmniController'
 
     def __init__(self, **kwargs) -> None:
         super().__init__(**kwargs)
-        self.channelType = PressureChannel
-        self.controller = PressureController(controllerParent=self)
+        self.channelType = OmniChannel
+        self.controller = OmniController(controllerParent=self)
 
-    def getChannels(self) -> 'list[PressureChannel]':
-        return cast('list[PressureChannel]', super().getChannels())
+    def getChannels(self) -> 'list[OmniChannel]':
+        return cast('list[OmniChannel]', super().getChannels())
 
     com: str
 
@@ -45,53 +52,218 @@ class OMNICONTROL(Device):
         return defaultSettings
 
 
-class PressureChannel(Channel):
+class OmniChannel(Channel):
     """UI for pressure with integrated functionality."""
 
-    ID = 'ID'
     channelParent: OMNICONTROL
+    ID = 'ID'
+    OMNITYPE = 'OMNITYPE'
+    OMNIICON = 'OMNIICON'
+    PumpStatn = 'PumpStatn'
+    Standby = 'Standby'
+    DrvPower = 'DrvPower'
+    TempPump = 'TempPump'
+    ERRORLED = 'ERRORLED'
+
+    class OMNITYPESTATE(Enum):
+        """States for the Stack Action."""
+
+        PRESSURE = 'Pressure Sensor'
+        TMP = 'Turbo Pump'
+
+    def initGUI(self, item: dict) -> None:
+        super().initGUI(item)
+        omniIcon = self.getParameterByName(self.OMNIICON)
+        omniIcon.widget = QPushButton()
+        omniIcon.widget.setStyleSheet('QPushButton{border:none;}')
+        omniIcon.applyWidget()
+
+        pumpStatn = self.getParameterByName(self.PumpStatn)
+        initialValue = pumpStatn.value or False
+        pumpStatn.widget = ToolButton()  # hard to spot checked QCheckBox. QPushButton is too wide -> overwrite internal widget to QToolButton
+        pumpStatn.applyWidget()
+        pumpStatn.widget.setMaximumHeight(pumpStatn.rowHeight)  # default too high
+        pumpStatn.widget.setText('ON')
+        pumpStatn.widget.setMinimumWidth(5)
+        pumpStatn.widget.setCheckable(True)
+        pumpStatn.value = initialValue
+
+        standby = self.getParameterByName(self.Standby)
+        initialValue = standby.value or False
+        standby.widget = ToolButton()  # hard to spot checked QCheckBox. QPushButton is too wide -> overwrite internal widget to QToolButton
+        standby.applyWidget()
+        standby.widget.setMaximumHeight(standby.rowHeight)  # default too high
+        standby.widget.setText(self.Standby.title())
+        standby.widget.setMinimumWidth(5)
+        standby.widget.setCheckable(True)
+        standby.value = initialValue
+
+        self.realChanged()
+        self.omniTypeChanged()
 
     def getDefaultChannel(self) -> dict[str, dict]:
 
         # definitions for type hinting
         self.id: int
+        self.omniType: str
+        self.pumpStatn: bool
+        self.standby: bool
+        self.drvPower: float
+        self.tempPump: float
+        self.notes: str
+        self.errorLED: bool
 
         channel = super().getDefaultChannel()
-        channel[self.VALUE][Parameter.HEADER] = 'P (mbar)'
+        channel[self.VALUE][Parameter.HEADER] = 'P (mbar) / S (Hz)'
         channel[self.ID] = parameterDict(value=1, parameterType=PARAMETERTYPE.INTCOMBO, advanced=True,
-                                        items='0, 1, 2, 3, 4, 5, 6', attr='id')
+                                        items='0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16', attr='id')
+        channel[self.OMNIICON] = parameterDict(value=False, parameterType=PARAMETERTYPE.BOOL, advanced=False,
+                                                 toolTip='Device type.', header=' ')
+        channel[self.OMNITYPE] = parameterDict(value=self.OMNITYPESTATE.PRESSURE.value, parameterType=PARAMETERTYPE.COMBO, advanced=True,
+                                               toolTip='Select which type of device this channel is controlling.',
+                                        items=f'{self.OMNITYPESTATE.PRESSURE.value}, {self.OMNITYPESTATE.TMP.value}', attr='omniType', header='Type', event=self.omniTypeChanged)
+        channel[self.PumpStatn] = parameterDict(value=False, parameterType=PARAMETERTYPE.BOOL, advanced=True,
+                                        attr='pumpStatn', restore=False, event=self.setPumpStatn, header='State')
+        channel[self.Standby] = parameterDict(value=False, parameterType=PARAMETERTYPE.BOOL, advanced=True,
+                                        attr='standby', restore=False, event=self.setStandby)
+        channel[self.DrvPower] = parameterDict(value=np.nan, parameterType=PARAMETERTYPE.FLOAT, advanced=True,
+                                        attr='drvPower', restore=False, header='Power (W)')
+        channel[self.TempPump] = parameterDict(value=np.nan, parameterType=PARAMETERTYPE.FLOAT, advanced=True,
+                                        attr='tempPump', restore=False, header='Temp (°C)')
+        channel[self.ERRORLED] = parameterDict(value=False, parameterType=PARAMETERTYPE.BOOL, advanced=True, indicator=True,
+                                        header='Err', toolTip='Indicates errors.', attr='errorLED', restore=False)
+        channel[self.NOTES] = parameterDict(value='', parameterType=PARAMETERTYPE.LABEL, advanced=True, attr='notes', restore=False, indicator=True)
         return channel
 
     def setDisplayedParameters(self) -> None:
         super().setDisplayedParameters()
+        self.insertDisplayedParameter(self.OMNIICON, before=self.ENABLED)
+        self.insertDisplayedParameter(self.OMNITYPE, before=self.ENABLED)
+        self.insertDisplayedParameter(self.PumpStatn, before=self.VALUE)
+        self.insertDisplayedParameter(self.Standby, before=self.VALUE)
+        self.insertDisplayedParameter(self.DrvPower, before=self.EQUATION)
+        self.insertDisplayedParameter(self.TempPump, before=self.EQUATION)
+        self.insertDisplayedParameter(self.ERRORLED, before=self.EQUATION)
         self.displayedParameters.append(self.ID)
+        self.displayedParameters.append(self.NOTES)
+
+    def tempParameters(self) -> list[str]:
+        return [*super().tempParameters(), self.OMNIICON]
+
+    def realChanged(self) -> None:
+        super().realChanged()
+        self.hideParameters()
+
+    def hideParameters(self) -> None:
+        self.getParameterByName(self.PumpStatn).setVisible(self.omniType == self.OMNITYPESTATE.TMP.value)
+        self.getParameterByName(self.Standby).setVisible(self.real and self.omniType == self.OMNITYPESTATE.TMP.value)
+        self.getParameterByName(self.DrvPower).setVisible(self.real and self.omniType == self.OMNITYPESTATE.TMP.value)
+        self.getParameterByName(self.TempPump).setVisible(self.real and self.omniType == self.OMNITYPESTATE.TMP.value)
+
+    def omniTypeChanged(self) -> None:
+        self.hideParameters()
+        omniIconPushButton = cast('QPushButton', self.getParameterByName(self.OMNIICON).getWidget())
+        omniIconPushButton.setIcon(self.channelParent.makeIcon('sensor.png' if self.omniType == self.OMNITYPESTATE.PRESSURE.value else 'turbo.png'))
+        self.getParameterByName(self.OMNIICON).toolTip = self.omniType
+        oldValue = self.value
+        value = self.getParameterByName(self.VALUE)
+        value.parameterType = PARAMETERTYPE.EXP if self.omniType == self.OMNITYPESTATE.PRESSURE.value else PARAMETERTYPE.INT
+        value.applyWidget()
+        self.scalingChanged()
+        self.updateColor()
+        self.value = oldValue
+        self.logY = self.omniType == self.OMNITYPESTATE.PRESSURE.value
+
+    @property
+    def unit(self) -> str:
+        """The unit depending on device type."""
+        return 'mBar' if self.omniType == self.OMNITYPESTATE.PRESSURE.value else 'Hz'
+
+    def checkPumpChannel(self) -> bool:
+        isPump = self.omniType == self.OMNITYPESTATE.TMP.value
+        if not isPump:
+            self.print(f'{self.name} is not a Turbo Pump.')
+        return isPump
+
+    def setPumpStatn(self) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.setPumpStatn(addr=self.id, on=self.pumpStatn)
+
+    def setStandby(self) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.setStandby(addr=self.id, on=self.standby)
+
+    def getStandby(self) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.getStandby(addr=self.id)
+
+    def getStdbySVal(self) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.getStdbySVal(addr=self.id)
+
+    def setStdbySVal(self, value: float) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.setStdbySVal(addr=self.id, value=value)
+
+    def acknError(self) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.acknError(addr=self.id)
+
+    def setRS485Adr(self, newAddr: int) -> None:
+        if not self.checkPumpChannel():
+            return
+        self.channelParent.controller.setRS485Adr(addr=self.id, newAddr=newAddr)
 
 
-class PressureController(DeviceController):
+class OmniController(DeviceController):  # noqa: PLR0904
 
     controllerParent: OMNICONTROL
 
     def runInitialization(self) -> None:
         try:
             self.port = serial.Serial(self.controllerParent.com, timeout=1)
-            pvp.enable_valid_char_filter()
             self.signalComm.initCompleteSignal.emit()
         except Exception as e:  # pylint: disable=[broad-except]  # noqa: BLE001
             self.print(f'Error while initializing: {e}', flag=PRINT.ERROR)
         finally:
             self.initializing = False
 
+    def initializeValues(self, reset: bool = False) -> None:
+        super().initializeValues(reset)
+        self.pumpStatn = [False] * len(self.controllerParent.getChannels())  # initializing values, overwrite if needed
+        self.standby = [False] * len(self.controllerParent.getChannels())  # initializing values, overwrite if needed
+        self.drvPower = np.array([np.nan] * len(self.controllerParent.getChannels()))  # initializing values, overwrite if needed
+        self.tempPump = np.array([np.nan] * len(self.controllerParent.getChannels()))  # initializing values, overwrite if needed
+        self.errorCode = [''] * len(self.controllerParent.getChannels())  # initializing values, overwrite if needed
+
     def readNumbers(self) -> None:
         for i, channel in enumerate(self.controllerParent.getChannels()):
             if channel.enabled and channel.active and channel.real:
                 try:
-                    pressure = pvp.read_pressure(self.port, channel.id)
-                    self.print(f'readNumbers channel.id: {channel.id}, response {pressure}', flag=PRINT.TRACE)
-                    self.values[i] = np.nan if pressure == 0 else pressure * 1000
+                    if channel.omniType == channel.OMNITYPESTATE.PRESSURE.value:
+                        pressure = self.getPressure(addr=channel.id, already_acquired=True)
+                        self.print(f'readNumbers channel.id: {channel.id}, response {pressure}', flag=PRINT.TRACE)
+                        self.values[i] = np.nan if pressure == 0 else pressure
+                    else:
+                        speed = self.getActualSpd(addr=channel.id, already_acquired=True)
+                        self.pumpStatn[i] = self.getPumpStatn(addr=channel.id, already_acquired=True)
+                        self.standby[i] = self.getStandby(addr=channel.id, already_acquired=True)
+                        self.drvPower[i] = self.getDrvPower(addr=channel.id, already_acquired=True)
+                        self.tempPump[i] = self.getTempPump(addr=channel.id, already_acquired=True)
+                        self.errorCode[i] = self.getErrorCode(addr=channel.id, already_acquired=True)
+                        self.values[i] = speed
                 except ValueError as e:
-                    self.print(f'Error while reading pressure {e}')
+                    self.print(f"Error while reading {'pressure' if channel.omniType == channel.OMNITYPESTATE.PRESSURE.value else 'speed'} {e}")
                     self.errorCount += 1
                     self.values[i] = np.nan
+                    self.drvPower[i] = np.nan
+                    self.tempPump[i] = np.nan
 
     def rndPressure(self) -> float:
         """Return a random pressure."""
@@ -102,7 +274,32 @@ class PressureController(DeviceController):
     def fakeNumbers(self) -> None:
         for i, channel in enumerate(self.controllerParent.getChannels()):
             if channel.enabled and channel.active and channel.real:
-                self.values[i] = self.rndPressure() if np.isnan(self.values[i]) else self.values[i] * self.rng.uniform(.99, 1.01)  # allow for small fluctuation
+                if channel.omniType == channel.OMNITYPESTATE.PRESSURE.value:
+                    self.values[i] = self.rndPressure() if np.isnan(self.values[i]) else self.values[i] * self.rng.uniform(.99, 1.01)  # allow for small fluctuation
+                else:
+                    self.values[i] = (0 if np.isnan(self.values[i]) else  # simulate acceleration and deceleration and allow for small fluctuation
+                    max(0, min(1500, (self.values[i] + (25 if channel.pumpStatn and (channel.value < (1000 if channel.standby else 1500)) else - 25) * self.controllerParent.interval / 1000))) +
+                    self.rng.uniform(-5, 5))
+                    self.pumpStatn[i] = channel.pumpStatn
+                    self.standby[i] = channel.standby
+                    self.drvPower[i] = self.values[i] / 30
+                    self.tempPump[i] = 25 + self.values[i] / 60
+                    self.errorCode[i] = '000000'
+
+    def updateValues(self) -> None:
+        super().updateValues()
+        for i, channel in enumerate(self.controllerParent.getChannels()):
+            if channel.enabled and channel.real and channel.omniType == channel.OMNITYPESTATE.TMP.value:
+                channel.pumpStatn = self.pumpStatn[i]
+                channel.standby = self.standby[i]
+                channel.drvPower = self.drvPower[i]
+                channel.tempPump = self.tempPump[i]
+                if self.errorCode[i] == '000000':
+                    channel.notes = 'No Error'
+                    channel.errorLED = False
+                else:
+                    channel.notes = self.errorCode[i]
+                    channel.errorLED = True
 
     def closeCommunication(self) -> None:
         super().closeCommunication()
@@ -112,3 +309,126 @@ class PressureController(DeviceController):
                 self.port = None
         self.initialized = False
         self.closing = False
+
+    def getPressure(self, addr: int, already_acquired=False) -> float:
+        """Get pressure from gauge in mBar."""
+        response = self.OmniWriteRead(addr=addr, param_num=740, already_acquired=already_acquired)
+        mantissa = int(response[:4])
+        exponent = int(response[4:])
+        return float(mantissa * 10 ** (exponent - 26)) * 1000
+
+    def getActualSpd(self, addr: int, already_acquired=False) -> int:
+        return int(self.OmniWriteRead(addr=addr, param_num=309, already_acquired=already_acquired))
+
+    def getDrvPower(self, addr: int, already_acquired=False) -> int:
+        return int(self.OmniWriteRead(addr=addr, param_num=316, already_acquired=already_acquired))
+
+    def getPumpStatn(self, addr: int, already_acquired=False) -> bool:
+        return self.OmniWriteRead(addr=addr, param_num=10, already_acquired=already_acquired) == '111111'
+
+    def setPumpStatn(self, addr: int, on: bool, already_acquired=False) -> bool:
+        return self.OmniWriteRead(addr=addr, param_num=10, value=111111 if on else 0, already_acquired=already_acquired) == '111111'
+
+    def getTempPump(self, addr: int, already_acquired=False) -> int:  # 384 TempRotor, 346 TempMotor, 330 TempPmpBot
+        return int(self.OmniWriteRead(addr=addr, param_num=330, already_acquired=already_acquired))
+
+    def getStandby(self, addr: int, already_acquired=False) -> bool:
+        return self.OmniWriteRead(addr=addr, param_num=2, already_acquired=already_acquired) == '111111'
+
+    def setStandby(self, addr: int, on: bool, already_acquired=False) -> bool:
+        return self.OmniWriteRead(addr=addr, param_num=2, value=111111 if on else 0, already_acquired=already_acquired) == '111111'
+
+    def getErrorCode(self, addr: int, already_acquired=False) -> str:
+        return self.OmniWriteRead(addr=addr, param_num=303, already_acquired=already_acquired)
+
+    def getStdbySVal(self, addr: int, already_acquired=False) -> float:
+        return float(self.OmniWriteRead(addr=addr, param_num=717, already_acquired=already_acquired)) / 100
+
+    def setStdbySVal(self, addr: int, value: float, already_acquired=False) -> float:
+        if value < 20 or value > 100:  # noqa: PLR2004
+            self.print(f'Standby speed value {value} invalid. Value needs to be between 20 and 100.')
+            return self.getStdbySVal(addr=addr)
+        return float(self.OmniWriteRead(addr=addr, param_num=717, value=value * 100, already_acquired=already_acquired)) / 100
+
+    def setRS485Adr(self, addr: int, newAddr: int, already_acquired=False) -> int:
+        return int(self.OmniWriteRead(addr=addr, param_num=797, value=newAddr, already_acquired=already_acquired))
+
+    def acknError(self, addr: int, already_acquired=False) -> None:
+        self.OmniWriteRead(addr=addr, param_num=9, value=1, already_acquired=already_acquired)
+
+    def makeOmniMessage(self, addr: int, param_num: int, value=None) -> str:
+        """Generate a message to query or set the specified parameter.
+
+        :param addr: the address of the device
+        :type addr: int
+        :param param_num: Parameter number
+        :type param_num: int
+        :param value: set value. If this is None, this returns a query, otherwise a set command.
+        :type value: int
+        :return: The query or set message.
+        :rtype: str
+        """
+        message = ''
+        if value is None:
+            message = f'{addr:03d}00{param_num:03d}02=?'
+            message += f'{sum(ord(x) for x in message) % 256:03d}\r'
+        else:
+            message = f'{addr:03d}10{param_num:03d}06{value:06d}'
+            message += f'{sum(ord(x) for x in message) % 256:03d}\r'
+        return message
+
+    def parseOmniResponse(self, message: str) -> str:
+
+        # Evaluate the checksum
+        if int(message[-3:]) != (sum(ord(x) for x in message[:-3]) % 256):
+            self.print(f'Invalid checksum in response {message}.', flag=PRINT.WARNING)
+
+        addr = int(message[:3])
+        rw = int(message[3:4])
+        param_num = int(message[5:8])
+        data = message[10:-3]
+        self.print(f'Parsed response {message}: addr {addr}, rw {rw}, param_num {param_num}, data {data}.', flag=PRINT.TRACE)
+        return data
+
+    def OmniWrite(self, message: str) -> None:
+        """Pfeiffer specific serial write.
+
+        :param message: The serial message to be send.
+        :type message: str
+        """
+        if self.port:
+            self.serialWrite(self.port, f'{message}\r', encoding='ascii')
+
+    def OmniRead(self) -> str:
+        """Pfeiffer specific serial read.
+
+        :return: The serial response received.
+        :rtype: str
+        """
+        if self.port:
+            return self.serialRead(self.port, EOL='\r', encoding='ascii')  # response
+        return ''
+
+    def OmniWriteRead(self, addr: int, param_num: int, value=None, already_acquired: bool = False) -> str:
+        """Pfeiffer specific serial write and read.
+
+        :param addr: the address of the device
+        :type addr: int
+        :param param_num: Parameter number
+        :type param_num: int
+        :param value: set value. If this is None, this returns a query, otherwise a set command.
+        :type value: int
+        :param already_acquired: Indicates if the lock has already been acquired, defaults to False
+        :type already_acquired: bool, optional
+        :return: The serial response received.
+        :rtype: str
+        """
+        response = ''
+        message = self.makeOmniMessage(addr=addr, param_num=param_num, value=value)
+        with self.lock.acquire_timeout(2, timeoutMessage=f'Cannot acquire lock for message: {message}', already_acquired=already_acquired) as lock_acquired:
+            if lock_acquired:
+                self.OmniWrite(message)
+                response = self.OmniRead()  # reads return value
+                if response:
+                    response = self.parseOmniResponse(response)
+        return response